Rename modules with leading underscores and update imports
Renamed 'interactive.py', 'line.py', and 'utilities.py' to '_interactive.py', '_line.py', and '_utilities.py' respectively to indicate internal use. Updated imports in '__init__.py' accordingly. Also changed version import to use '_version' module with a fallback to 'unknown' if unavailable.
