+++ src/napari_signal_selector/__init__.py
@@ -1,6 +1,11 @@
-__version__ = "0.0.6"
+try:
+    from ._version import __version__
+except ImportError:
+    __version__ = "unknown"
+
+
 from ._sample_data import load_flashing_polygons_data, load_blinking_polygons_data
-from .interactive import InteractiveFeaturesLineWidget
+from ._interactive import InteractiveFeaturesLineWidget
 
 __all__ = (
     "InteractiveFeaturesLineWidget",
