+++ src/napari_signal_selector/_interactive.py
