+++ src/napari_signal_selector/_line.py
