+++ src/napari_signal_selector/_utilities.py
